tests/wpt/tests/css/css-text/parsing/word-space-transform-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().wordSpaceTransform</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-space-transform">
<meta name="assert" content="word-space-transform computed value is specified keyword.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("word-space-transform", "none");
test_computed_value("word-space-transform", "space");
test_computed_value("word-space-transform", "ideographic-space");
test_computed_value("word-space-transform", "space auto-phrase");
test_computed_value("word-space-transform", "ideographic-space auto-phrase");
</script>
</body>
</html>
